progress on vehicle movement, intersections connect with each other, will finish rest of turning cases later

+++ src/vector.c
@@ -1,6 +1,6 @@
 #include "prototypes.h"
 
-void init_vector(vector_T* vector, short x, short y)
+void init_vector(vector_T* vector, int x, int y)
 {
     vector->x = x;
     vector->y = y;
